REN-91
Added CRUD endpoints for a company to create a building, add condos to a building, and to view all buildings, all condos per building, and the count of units per building. There are three TODO endpoints for updating a building, updating a condo, and retrieving information about a condo unit using buildingId and condoId. Considering that all IDs are unique across the database, we might remove certain IDs from the endpoint URLs. Also, added comprehensive exception handling for these endpoints.